DEDUCE/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deduce"
version = "0.1.0"
description = "A configurable framework for zero-shot semantic prediction using multi-modal embeddings"
readme = "README.rst"
license = {text = "MIT"}
authors = [
    {name = "Connor Hashemi", email = "connor.hashemi@kitware.com"}
]
dependencies = [
    "annotated-types==0.7.0",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.7",
    "clean-fid==0.1.35",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "cuda-bindings==13.2.0",
    "cuda-pathfinder==1.5.3",
    "cuda-toolkit==13.0.2",
    "cycler==0.12.1",
    "dataeval==1.0.6",
    "et-xmlfile==2.0.0",
    "filelock==3.28.0",
    "fonttools==4.62.1",
    "fsspec==2026.3.0",
    "idna==3.11",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "kiwisolver==1.5.0",
    "lightgbm==4.6.0",
    "llvmlite==0.47.0",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numba==0.65.0",
    "numpy==2.2.6",
    "nvidia-cublas==13.1.0.3",
    "nvidia-cuda-cupti==13.0.85",
    "nvidia-cuda-nvrtc==13.0.88",
    "nvidia-cuda-runtime==13.0.96",
    "nvidia-cudnn-cu13==9.19.0.56",
    "nvidia-cufft==12.0.0.61",
    "nvidia-cufile==1.15.1.6",
    "nvidia-curand==10.4.0.35",
    "nvidia-cusolver==12.0.4.66",
    "nvidia-cusparse==12.6.3.3",
    "nvidia-cusparselt-cu13==0.8.0",
    "nvidia-nccl-cu13==2.28.9",
    "nvidia-nvjitlink==13.0.88",
    "nvidia-nvshmem-cu13==3.4.5",
    "nvidia-nvtx==13.0.85",
    "open-clip-torch>=3.3.0",
    "openpyxl==3.1.5",
    "packaging==26.1",
    "pandas==2.3.3",
    "pillow==12.2.0",
    "polars==1.39.3",
    "polars-runtime-32==1.39.3",
    "psutil==7.2.2",
    "pydantic==2.13.0",
    "pydantic-core==2.46.0",
    "pyparsing==3.3.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2026.1.post1",
    "requests==2.33.1",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "setuptools==81.0.0",
    "six==1.17.0",
    "sympy==1.14.0",
    "tabulate==0.10.0",
    "threadpoolctl==3.6.0",
    "torch==2.11.0",
    "torchvision==0.26.0",
    "tqdm==4.67.3",
    "triton==3.6.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2026.1",
    "urllib3==2.6.3",
    "xxhash==3.6.0",
]

[project.optional-dependencies]

[project.urls]
Repository = "https://github.com/Hashe037/DEDUCE"

[tool.setuptools.packages.find]
where = ["src"]
include = ["deduce*"]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
